Extract desktop demo renderer module
